🐛 fix	group탈퇴 및 추방 시 속하지 않은 이벤트와 일정에 대한 탈퇴 로직 제외
현재 나가는 대상의 member를 포함하지 않은 event와 schedule에 대해서 내보내는 로직이 돌아, 포함되지 않은 schedule에서 그룹장인지 확인 하는 등 불필요한 로직으로 예상치 못한 예외가 발생 중입니다.

+++ src/main/java/com/grepp/spring/app/model/group/service/GroupCommandService.java
@@ -369,6 +369,9 @@ public class GroupCommandService {
     private void deportMemberInEvent(GroupMember targetGroupMember, Long groupId) {
         for (Event event : eventRepository.findByGroupId(groupId)) {
             // eventMember 삭제
+            if(!eventMemberRepository.existsByEventIdAndMemberId(event.getId(), targetGroupMember.getMember().getId())){
+                continue;
+            }
             eventMemberRepository.deleteByEventAndMemberId(event,
                 targetGroupMember.getMember().getId());
             // schedule에서 추방
@@ -386,6 +389,9 @@ public class GroupCommandService {
     private void deportMemberInSchedule(GroupMember targetGroupMember, Event event) {
         // 이벤트 내 일정에 대한 처리
         for (Schedule schedule : scheduleCommandRepository.findByEvent(event)) {
+            if(!scheduleMemberCommandRepository.existsByScheduleIdAndMemberId(schedule.getId(), targetGroupMember.getMember().getId())){
+                continue;
+            }
             // scheduleMember 삭제
             // 일정에 본인이 일정 팀장인 경우 isRoleMaster를 true로 설정
             boolean isRoleMaster = false;
